requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.16.5
altair==5.2.0
anyio==3.7.1
applicationinsights==0.11.10
appnope==0.1.3
arrow==1.2.3
astor==0.8.1
asttokens==2.4.0
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.1
av==10.0.0
azure-ai-formrecognizer==3.3.0
azure-cognitiveservices-speech==1.30.0
azure-common==1.1.28
azure-core==1.29.0
azure-functions==1.15.0
azure-storage-blob==12.17.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bidict==0.22.1
bleach==6.0.0
blessed==1.20.0
blinker==1.6.2
boto3==1.33.1
botocore==1.33.1
braceexpand==0.1.7
Brotli==1.1.0
cachelib==0.10.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
coloredlogs==15.0.1
contourpy==1.1.1
croniter==1.3.15
cryptography==41.0.4
ctranslate2==3.16.1
cycler==0.12.1
dataclasses-json==0.5.7
dateparser==1.1.8
dateutils==0.6.12
decorator==4.4.2
deepdiff==6.7.1
distlib==0.3.7
dlib==19.24.2
dnspython==2.4.1
docopt==0.6.2
dominate==2.8.0
duckdb==0.8.1
et-xmlfile==1.1.0
eventlet==0.33.3
executing==1.2.0
fastapi==0.88.0
ffmpeg==1.4
ffmpy==0.3.1
ffprobe-python==1.0.3
filelock==3.12.4
Flask==2.3.2
flask-babel==3.1.0
Flask-BabelEx==0.9.4
Flask-Bootstrap==3.3.7.1
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-Migrate==4.1.0
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.2
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
ftfy==6.1.1
gdown==4.7.1
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.4.1
google-cloud-storage==2.15.0
google-cloud-translate==3.15.3
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.59.1
googletrans==4.0.0rc1
gradio==3.47.1
gradio_client==0.6.0
greenlet==2.0.2
grpcio==1.62.1
grpcio-status==1.62.1
gTTS==2.3.2
gunicorn==21.2.0
h11==0.9.0
h2==3.2.0
h5py==3.8.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
huggingface-hub==0.19.4
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-resources==6.1.1
inquirer==3.1.4
install==1.3.5
ipython==8.14.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpickle==3.0.2
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kiwisolver==1.4.5
langchain==0.0.306
langdetect==1.0.9
langsmith==0.0.38
lazy_loader==0.3
librosa==0.10.1
lightning==2.0.0
lightning-cloud==0.5.55
lightning-utilities==0.10.0
llama-index==0.8.39.post2
llvmlite==0.40.1
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-enum==1.5.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
moment==0.12.1
more-itertools==9.1.0
moviepy==1.0.3
mpmath==1.3.0
msal==1.22.0
msgpack==1.0.7
msrest==0.7.1
multidict==6.0.4
mutagen==1.47.0
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
nest-asyncio==1.5.8
netifaces==0.10.6
networkx==3.1
neuralspace==1.0.6
nltk==3.8.1
numba==0.57.0
numexpr==2.8.7
numpy==1.24.3
oauthlib==3.2.2
onnxruntime==1.15.1
openai==0.27.8
openai-whisper @ git+https://github.com/openai/whisper.git@248b6cb124225dd263bb9bd32d060b6517e067f8
openapi-schema-pydantic==1.2.4
opencv-python==4.8.0.74
openpyxl==3.1.2
ordered-set==4.1.0
orjson==3.9.10
packaging==23.1
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
platformdirs==3.10.0
pooch==1.8.0
proglog==0.1.10
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.21.12
psutil==5.9.6
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11 @ file:///private/tmp/pybind11-20231006-5480-1dj4kz4/pybind11-2.11.1
pycparser==2.21
pycryptodomex==3.18.0
pydantic==1.10.9
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.7.0
pyodbc==4.0.39
pyparsing==3.1.0
pypdf==3.9.1
PyPDF2==3.0.1
pypyodbc==1.3.6
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-engineio==4.4.1
python-multipart==0.0.6
pytorch-lightning==2.1.2
pytube==15.0.0
pytz==2023.3
pyvis==0.3.2
PyYAML==6.0
readchar==4.0.5
referencing==0.31.0
regex==2023.6.3
requests==2.29.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.7.0
rpds-py==0.13.1
rsa==4.9
s3transfer==0.8.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
semantic-version==2.10.0
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
speaklater==1.3
SQLAlchemy==2.0.16
stack-data==0.6.2
starlette==0.22.0
starsessions==1.3.0
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.2.0
tiktoken==0.3.3
times==0.7
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1
torchlibrosa==0.1.0
torchmetrics==1.2.0
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.1
typer==0.9.0
types-python-dateutil==2.8.19.14
typing-inspect==0.8.0
typing_extensions==4.15.0
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.24.0.post1
virtualenv==20.23.1
visitor==0.1.3
watchdog==3.0.0
wcwidth==0.2.6
webdataset==0.2.48
webencodings==0.5.1
websocket-client==1.6.4
websockets==11.0.3
Werkzeug==2.3.6
wget==3.2
wsproto==1.2.0
WTForms==3.2.1
yarg==0.1.9
yarl==1.9.2
youtube-dl==2021.12.17
youtube-transcript-api==0.6.1
yt-dlp==2023.7.6
